clear up stupid gcc issues

+++ net_io.c
@@ -2259,7 +2259,7 @@ struct char_buffer generateGlobeBin(int globe_index){
     uint32_t ac_count_pos = Modes.json_ac_count_pos;
     memWrite(p, ac_count_pos);
 
-    if (p - buf > elementSize)
+    if (p - buf > (int) elementSize)
         fprintf(stderr, "buffer overrun globeBin\n");
 
     p = buf + elementSize;
